requirements.txt
discord.py
python-dotenv
numpy
aiosqlite
